job-finder-api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-cloud-storage==2.14.0
PyPDF2==3.0.1
sentence-transformers==2.2.2
torch==2.1.0
python-multipart==0.0.6